Split resolver metadata tests
